working hmr
